@@ -25,14 +25,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.2",
-    "@angular/common": "^13.2.2",
-    "@angular/compiler": "^13.2.2",
-    "@angular/core": "^13.2.2",
-    "@angular/forms": "^13.2.2",
-    "@angular/platform-browser": "^13.2.2",
-    "@angular/platform-browser-dynamic": "^13.2.2",
-    "@angular/router": "^13.2.2",
+    "@angular/animations": "^14.0.4",
+    "@angular/common": "^14.0.4",
+    "@angular/compiler": "^14.0.4",
+    "@angular/core": "^14.0.4",
+    "@angular/forms": "^14.0.4",
+    "@angular/platform-browser": "^14.0.4",
+    "@angular/platform-browser-dynamic": "^14.0.4",
+    "@angular/router": "^14.0.4",
     "bugsplat": "^7.1.0",
     "core-js": "^2.6.3",
     "rxjs": "^6.6.3",
@@ -40,16 +40,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.3",
-    "@angular-devkit/build-ng-packagr": "^0.1002.0",
-    "@angular-eslint/builder": "13.0.1",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/schematics": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "^13.2.3",
-    "@angular/compiler-cli": "^13.2.2",
-    "@angular/language-service": "^13.2.2",
+    "@angular-devkit/build-angular": "~14.0.4",
+    "@angular-eslint/builder": "14.0.0",
+    "@angular-eslint/eslint-plugin": "14.0.0",
+    "@angular-eslint/eslint-plugin-template": "14.0.0",
+    "@angular-eslint/schematics": "14.0.0",
+    "@angular-eslint/template-parser": "14.0.0",
+    "@angular/cli": "^14.0.4",
+    "@angular/compiler-cli": "^14.0.4",
+    "@angular/language-service": "^14.0.4",
     "@bugsplat/symbol-upload": "^4.1.1",
     "@commitlint/cli": "^17.0.3",
     "@commitlint/config-conventional": "^17.0.3",
@@ -58,14 +57,17 @@
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "^12.20.55",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
     "copyfiles": "^2.4.1",
     "dotenv": "^10.0.0",
-    "eslint": "^8.2.0",
+    "eslint": "^8.18.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsdoc": "^39.3.3",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
     "http-server": "^0.12.3",
     "husky": "^8.0.1",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "^3.99.1",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
@@ -73,12 +75,11 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "karma-teamcity-reporter": "^1.1.0",
-    "ng-packagr": "^13.2.1",
-    "protractor": "~7.0.0",
+    "ng-packagr": "^14.0.2",
+    "semantic-release": "^19.0.3",
     "semantic-release-cli": "^5.4.4",
     "ts-node": "~8.0.2",
-    "typescript": "~4.5.5",
-    "semantic-release": "^19.0.3"
+    "typescript": "~4.7.4"
   },
   "repository": {
     "type": "git",
