@@ -1,49 +1,49 @@
 
 export enum BugSplatLogLevel {
-    None,
-    Error,
-    Warn,
-    Info,
-    Log
+    none,
+    error,
+    warn,
+    info,
+    log
 }
 
 export interface Logger {
-    error(...params: (string | Object)[]): void;
-    warn(...params: (string | Object)[]): void;
-    info(...params: (string | Object)[]): void;
-    log(...params: (string | Object)[]): void;
+    error(...params: (string | object)[]): void;
+    warn(...params: (string | object)[]): void;
+    info(...params: (string | object)[]): void;
+    log(...params: (string | object)[]): void;
 }
 
 export class BugSplatLogger {
 
-    static readonly LEVEL_CANNOT_BE_NULL = "BugSplatLogger Error: level cannot be null or undefined!";
-    static readonly LOGGER_CANNOT_BE_NULL = "BugSplatLogger Error: logger cannot be null or undefined!";
+    static readonly levelCannotBeNull = 'BugSplatLogger Error: level cannot be null or undefined!';
+    static readonly loggerCannotBeNull = 'BugSplatLogger Error: logger cannot be null or undefined!';
 
-    constructor(private level: BugSplatLogLevel = BugSplatLogLevel.Error, private logger: Logger = console) {
+    constructor(private level: BugSplatLogLevel = BugSplatLogLevel.error, private logger: Logger = console) {
         if (this.level === null || this.level === undefined) {
-            throw new Error(BugSplatLogger.LEVEL_CANNOT_BE_NULL)
+            throw new Error(BugSplatLogger.levelCannotBeNull)
         }
         if (!this.logger) {
-            throw new Error(BugSplatLogger.LOGGER_CANNOT_BE_NULL);
+            throw new Error(BugSplatLogger.loggerCannotBeNull);
         }
     }
-    error(...params: (string | Object)[]): void {
-        if (this.level >= BugSplatLogLevel.Error) {
+    error(...params: (string | object)[]): void {
+        if (this.level >= BugSplatLogLevel.error) {
             this.logger.error(...params);
         }
     }
-    warn(...params: (string | Object)[]): void {
-        if (this.level >= BugSplatLogLevel.Warn) {
+    warn(...params: (string | object)[]): void {
+        if (this.level >= BugSplatLogLevel.warn) {
             this.logger.warn(...params);
         }
     }
-    info(...params: (string | Object)[]): void {
-        if (this.level >= BugSplatLogLevel.Info) {
+    info(...params: (string | object)[]): void {
+        if (this.level >= BugSplatLogLevel.info) {
             this.logger.info(...params);
         }
     }
-    log(...params: (string | Object)[]): void {
-        if (this.level >= BugSplatLogLevel.Log) {
+    log(...params: (string | object)[]): void {
+        if (this.level >= BugSplatLogLevel.log) {
             this.logger.log(...params);
         }
     }
