@@ -1,21 +1,21 @@
-import { BugSplatResponseData } from "./bugsplat-response-data";
+import { BugSplatResponseData } from './bugsplat-response-data';
 
 export enum BugSplatPostEventType {
-    Success,
-    Error
+    success,
+    error
 }
 
 export class BugSplatPostEvent {
 
-    static readonly TYPE_CANNOT_BE_NULL = "BugSplatPostEvent Error: type cannot be null, undefined, or empty";
-    static readonly RESPONSE_DATA_CANNOT_BE_NULL = "BugSplatPostEvent Error: responseData cannot be null, undefined, or empty";
+    static readonly typeCannotBeNull = 'BugSplatPostEvent Error: type cannot be null, undefined, or empty';
+    static readonly responseDataCannotBeNull = 'BugSplatPostEvent Error: responseData cannot be null, undefined, or empty';
 
     constructor(public type: BugSplatPostEventType, public responseData: BugSplatResponseData) {
         if (this.type === null || this.type === undefined) {
-            throw new Error(BugSplatPostEvent.TYPE_CANNOT_BE_NULL);
+            throw new Error(BugSplatPostEvent.typeCannotBeNull);
         }
         if (!this.responseData) {
-            throw new Error(BugSplatPostEvent.RESPONSE_DATA_CANNOT_BE_NULL);
+            throw new Error(BugSplatPostEvent.responseDataCannotBeNull);
         }
     }
 }
