@@ -3,21 +3,21 @@ import { BugSplatLogger, BugSplatLogLevel, Logger } from '../lib/bugsplat-logger
 describe('BugSplatLogger', () => {
 
     it('should throw if level is null', () => {
-        expect(() => new BugSplatLogger(null as any)).toThrowError(BugSplatLogger.LEVEL_CANNOT_BE_NULL);
+        expect(() => new BugSplatLogger(null as any)).toThrowError(BugSplatLogger.levelCannotBeNull);
     });
 
     it('should throw if logger is null', () => {
-        expect(() => new BugSplatLogger(BugSplatLogLevel.Info, null as any)).toThrowError(BugSplatLogger.LOGGER_CANNOT_BE_NULL);
+        expect(() => new BugSplatLogger(BugSplatLogLevel.info, null as any)).toThrowError(BugSplatLogger.loggerCannotBeNull);
     });
 
     it('should log errors when log level is set to Error', () => {
         const fakeLogger = createFakeLogger();
-        const errorSpy = spyOn(fakeLogger, "error");
-        const warnSpy = spyOn(fakeLogger, "warn");
-        const infoSpy = spyOn(fakeLogger, "info");
-        const logSpy = spyOn(fakeLogger, "log");
-        const expectedMessage = "BugSplat rocks!";
-        const sut = new BugSplatLogger(BugSplatLogLevel.Error, fakeLogger);
+        const errorSpy = spyOn(fakeLogger, 'error');
+        const warnSpy = spyOn(fakeLogger, 'warn');
+        const infoSpy = spyOn(fakeLogger, 'info');
+        const logSpy = spyOn(fakeLogger, 'log');
+        const expectedMessage = 'BugSplat rocks!';
+        const sut = new BugSplatLogger(BugSplatLogLevel.error, fakeLogger);
         sut.error(expectedMessage);
         sut.warn(expectedMessage);
         sut.info(expectedMessage);
@@ -30,12 +30,12 @@ describe('BugSplatLogger', () => {
 
     it('should log errors and warnings when log level is set to Warn', () => {
         const fakeLogger = createFakeLogger();
-        const errorSpy = spyOn(fakeLogger, "error");
-        const warnSpy = spyOn(fakeLogger, "warn");
-        const infoSpy = spyOn(fakeLogger, "info");
-        const logSpy = spyOn(fakeLogger, "log");
-        const expectedMessage = "BugSplat rocks!";
-        const sut = new BugSplatLogger(BugSplatLogLevel.Warn, fakeLogger);
+        const errorSpy = spyOn(fakeLogger, 'error');
+        const warnSpy = spyOn(fakeLogger, 'warn');
+        const infoSpy = spyOn(fakeLogger, 'info');
+        const logSpy = spyOn(fakeLogger, 'log');
+        const expectedMessage = 'BugSplat rocks!';
+        const sut = new BugSplatLogger(BugSplatLogLevel.warn, fakeLogger);
         sut.error(expectedMessage);
         sut.warn(expectedMessage);
         sut.info(expectedMessage);
@@ -48,12 +48,12 @@ describe('BugSplatLogger', () => {
 
     it('should log errors, warnings and info when log level is set to Info', () => {
         const fakeLogger = createFakeLogger();
-        const errorSpy = spyOn(fakeLogger, "error");
-        const warnSpy = spyOn(fakeLogger, "warn");
-        const infoSpy = spyOn(fakeLogger, "info");
-        const logSpy = spyOn(fakeLogger, "log");
-        const expectedMessage = "BugSplat rocks!";
-        const sut = new BugSplatLogger(BugSplatLogLevel.Info, fakeLogger);
+        const errorSpy = spyOn(fakeLogger, 'error');
+        const warnSpy = spyOn(fakeLogger, 'warn');
+        const infoSpy = spyOn(fakeLogger, 'info');
+        const logSpy = spyOn(fakeLogger, 'log');
+        const expectedMessage = 'BugSplat rocks!';
+        const sut = new BugSplatLogger(BugSplatLogLevel.info, fakeLogger);
         sut.error(expectedMessage);
         sut.warn(expectedMessage);
         sut.info(expectedMessage);
@@ -66,12 +66,12 @@ describe('BugSplatLogger', () => {
 
     it('should log errors, warnings, info and log when log level is set to Log', () => {
         const fakeLogger = createFakeLogger();
-        const errorSpy = spyOn(fakeLogger, "error");
-        const warnSpy = spyOn(fakeLogger, "warn");
-        const infoSpy = spyOn(fakeLogger, "info");
-        const logSpy = spyOn(fakeLogger, "log");
-        const expectedMessage = "BugSplat rocks!";
-        const sut = new BugSplatLogger(BugSplatLogLevel.Log, fakeLogger);
+        const errorSpy = spyOn(fakeLogger, 'error');
+        const warnSpy = spyOn(fakeLogger, 'warn');
+        const infoSpy = spyOn(fakeLogger, 'info');
+        const logSpy = spyOn(fakeLogger, 'log');
+        const expectedMessage = 'BugSplat rocks!';
+        const sut = new BugSplatLogger(BugSplatLogLevel.log, fakeLogger);
         sut.error(expectedMessage);
         sut.warn(expectedMessage);
         sut.info(expectedMessage);
@@ -84,12 +84,12 @@ describe('BugSplatLogger', () => {
 
     it('should not log when log level is set to None', () => {
         const fakeLogger = createFakeLogger();
-        const errorSpy = spyOn(fakeLogger, "error");
-        const warnSpy = spyOn(fakeLogger, "warn");
-        const infoSpy = spyOn(fakeLogger, "info");
-        const logSpy = spyOn(fakeLogger, "log");
-        const expectedMessage = "BugSplat rocks!";
-        const sut = new BugSplatLogger(BugSplatLogLevel.None, fakeLogger);
+        const errorSpy = spyOn(fakeLogger, 'error');
+        const warnSpy = spyOn(fakeLogger, 'warn');
+        const infoSpy = spyOn(fakeLogger, 'info');
+        const logSpy = spyOn(fakeLogger, 'log');
+        const expectedMessage = 'BugSplat rocks!';
+        const sut = new BugSplatLogger(BugSplatLogLevel.none, fakeLogger);
         sut.error(expectedMessage);
         sut.warn(expectedMessage);
         sut.info(expectedMessage);
@@ -101,19 +101,19 @@ describe('BugSplatLogger', () => {
     });
 
     it('should use console for logging if no logger is provided', () => {
-        const consoleSpy = spyOn(console, "log");
-        const expectedMessage = "BugSplat rocks!";
-        const sut = new BugSplatLogger(BugSplatLogLevel.Log);
+        const consoleSpy = spyOn(console, 'log');
+        const expectedMessage = 'BugSplat rocks!';
+        const sut = new BugSplatLogger(BugSplatLogLevel.log);
         sut.log(expectedMessage);
         expect(consoleSpy).toHaveBeenCalledWith(expectedMessage);
     });
-
-    function createFakeLogger(): Logger {
-        return {
-            error: (msg) => { },
-            warn: (msg) => { },
-            info: (msg) => { },
-            log: (msg) => { }
-        }
-    }
 });
+
+const createFakeLogger: () => Logger = () => {
+    return {
+        error: (msg) => { },
+        warn: (msg) => { },
+        info: (msg) => { },
+        log: (msg) => { }
+    }
+}
