@@ -8,11 +8,13 @@ describe('AppComponent', () => {
   let fixture: ComponentFixture<AppComponent>;
 
   beforeEach(() => {
-    let errorHandler = {
+    const responseData = {} as Record<string, number>;
+    responseData['crash_id'] = 99;
+    const errorHandler = {
       bugsplat: {
         database: 'fred',
         files: [],
-        getObservable: () => of({ responseData: { crash_id: 99 } }),
+        getObservable: () => of({ responseData }),
       },
     };
 
@@ -38,14 +40,11 @@ describe('AppComponent', () => {
   });
 
   it(`should have as title 'my-angular-crasher'`, () => {
-    const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
     expect(app.title).toEqual('my-angular-crasher');
   });
 
   it('should render title in a h1 tag', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
     expect(compiled.querySelector('h1').textContent).toContain(
       'Welcome to my-angular-crasher!'
